mariam_kipshidze_lecture/my_online_store/requirements.txt
Django==5.1
django-admin-sortable2==2.2.2
django-import-export==4.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
Markdown==3.7
